// src/generated/models/DashboardDailySnapshotModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */

export interface SafetyStatusValue {
  "@odata.type": string;
  Value: string;
  Id: number;
}

export interface QualityStatusValue {
  "@odata.type": string;
  Value: string;
  Id: number;
}

export interface ProcessDeviationStatusValue {
  "@odata.type": string;
  Value: string;
  Id: number;
}

export interface ProductDeviationStatusValue {
  "@odata.type": string;
  Value: string;
  Id: number;
}

export interface ProductivityStatusValue {
  "@odata.type": string;
  Value: string;
  Id: number;
}

export interface AuthorValue {
  "@odata.type": string;
  Claims: string;
  DisplayName: string;
  Email: string;
  Picture: string;
  Department: string;
  JobTitle: string;
}

export interface EditorValue {
  "@odata.type": string;
  Claims: string;
  DisplayName: string;
  Email: string;
  Picture: string;
  Department: string;
  JobTitle: string;
}

export interface DashboardDailySnapshotBase {
  ID?: number;
  Title?: string;
  ProductionDate?: string;
  "SafetyStatus#Id"?: number;
  "QualityStatus#Id"?: number;
  "ProcessDeviationStatus#Id"?: number;
  "ProductDeviationStatus#Id"?: number;
  "ProductivityStatus#Id"?: number;
  WeelyScrapPercent?: number;
  ProcessDeviationCount?: number;
  ProcessAffectedAreas?: number;
  ProductDeviationCount?: number;
  ProductAffectedAreas?: number;
  BuyoffCount?: number;
  MasterCardAvailabe?: number;
  MasterCardTotal?: number;
  MasterCardAvailabilityPercent?: number;
  ReproCompleted?: number;
  ReproPercent?: number;
  RepeatedRejectionCount?: number;
  OpenEscalationCount?: number;
  LongTermActionCount?: number;
  LastCalculated?: string;
  OData__ColorTag?: string;
  ComplianceAssetId?: string;
  Modified?: string;
  Created?: string;
  "Author#Claims"?: string;
  "Editor#Claims"?: string;
  "{Identifier}"?: string;
  "{IsFolder}"?: boolean;
  "{Thumbnail}"?: {
    "@odata.type"?: string;
    // URL to the small, highly compressed size thumbnail of the item, if available.
    Small?: string;
    // URL to the standard size thumbnail of the item, if available.
    Medium?: string;
    // URL to the largest size thumbnail of the item, if available.
    Large?: string;
  };
  "{Link}"?: string;
  "{Name}"?: string;
  "{FilenameWithExtension}"?: string;
  "{Path}"?: string;
  "{FullPath}"?: string;
  "{ModerationStatus}"?: string;
  "{ModerationComment}"?: string;
  "{ContentType}#Id"?: string;
  "{ContentType}"?: {
    "@odata.type"?: string;
    Id?: string;
    Name?: string;
  };
  "{HasAttachments}"?: boolean;
  "{Attachments}@odata.type"?: string;
  "{Attachments}"?: {
   "@odata.type"?: string;
   Id?: string;
   AbsoluteUri?: string;
   DisplayName?: string;
 }[];
  "{VersionNumber}"?: string;
  "{TriggerWindowStartToken}"?: string;
  "{TriggerWindowEndToken}"?: string;
}

export interface DashboardDailySnapshotRead extends DashboardDailySnapshotBase {
  SafetyStatus?: SafetyStatusValue;
  QualityStatus?: QualityStatusValue;
  ProcessDeviationStatus?: ProcessDeviationStatusValue;
  ProductDeviationStatus?: ProductDeviationStatusValue;
  ProductivityStatus?: ProductivityStatusValue;
  Author?: AuthorValue;
  Editor?: EditorValue;
}

export interface DashboardDailySnapshotWrite extends DashboardDailySnapshotBase {
  SafetyStatus?: string;
  QualityStatus?: string;
  ProcessDeviationStatus?: string;
  ProductDeviationStatus?: string;
  ProductivityStatus?: string;
  Author?: AuthorValue;
  Editor?: EditorValue;
}